Updated stopping criterion recommended for PASHA (#566)
The previous one is misleading for a lot of folks which are not very familiar with the library. They try out the code setting a small number of evaluations to see if it runs and this often leads to a failure because the number of configuration started is smaller than the number of workers.