feat: Backoff decorator for jobs scheduling (#673)
Simple backoff decorator which can be used for scheduling sagemaker jobs. Once a function is decorated with it, it will run normally but wait on a specific error (e.g. ResourceShareLimitExceededException) and retry after some time. This can be used to leave a scheduling loop running in the background and have it allocate new jobs when instances become available.